backend/requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch
tokenizers
sentence-transformers
numpy
faiss-cpu
httpx[http2]
kafka-python
pydantic-settings
accelerate